@@ -29,16 +29,16 @@ module HamcampsApp
     # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
     # config.i18n.default_locale = :de
     config.paperclip_defaults = {
-      :storage => :s3,
-      :region => 'us-west-1',
-      :s3_region => 'us-west-1',
+      storage: :s3,
+      region: 'us-west-1',
+      s3_region: 'us-west-1',
       s3_host_name: 's3-us-west-1.amazonaws.com',
-      :bucket => 'hamcamp',
-      :s3_credentials => {
-        :bucket => ENV["s3_bucket"],
-        :access_key_id => ENV["s3_access_key_id"],
-        :secret_access_key => ENV["s3_secret_access_key"],
-        :s3_region => ENV["s3_region"]
+      bucket: 'hamcamp',
+      s3_credentials: {
+        bucket: 'hamcamp',
+        access_key_id: 'AKIAJXCFZOWNWRMPZSVA',
+        secret_access_key: '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',
+        s3_region: 'us-west-1'
         }
       }
 
