@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -26,11 +25,10 @@ ActiveRecord::Schema.define(version: 20170119185915) do
     t.string   "image_content_type"
     t.integer  "image_file_size"
     t.datetime "image_updated_at"
+    t.index ["artist_id"], name: "index_albums_on_artist_id", using: :btree
+    t.index ["description"], name: "index_albums_on_description", using: :btree
+    t.index ["title"], name: "index_albums_on_title", using: :btree
   end
-
-  add_index "albums", ["artist_id"], name: "index_albums_on_artist_id", using: :btree
-  add_index "albums", ["description"], name: "index_albums_on_description", using: :btree
-  add_index "albums", ["title"], name: "index_albums_on_title", using: :btree
 
   create_table "tracks", force: :cascade do |t|
     t.string   "name",                   null: false
@@ -43,10 +41,9 @@ ActiveRecord::Schema.define(version: 20170119185915) do
     t.string   "track_url_content_type"
     t.integer  "track_url_file_size"
     t.datetime "track_url_updated_at"
+    t.index ["album_id"], name: "index_tracks_on_album_id", using: :btree
+    t.index ["name"], name: "index_tracks_on_name", using: :btree
   end
-
-  add_index "tracks", ["album_id"], name: "index_tracks_on_album_id", using: :btree
-  add_index "tracks", ["name"], name: "index_tracks_on_name", using: :btree
 
   create_table "users", force: :cascade do |t|
     t.string   "username",                   null: false
@@ -66,8 +63,7 @@ ActiveRecord::Schema.define(version: 20170119185915) do
     t.string   "profile_image_content_type"
     t.integer  "profile_image_file_size"
     t.datetime "profile_image_updated_at"
+    t.index ["username"], name: "index_users_on_username", using: :btree
   end
-
-  add_index "users", ["username"], name: "index_users_on_username", using: :btree
 
 end
