@@ -16,19 +16,19 @@ class MainHeader extends React.Component {
   componentDidMount() {
     // console.log(this.props);
 
-    let user = {
-        "key": "aa0ceb",
-        "firstName": this.props.user.username
-    };
+    // let user = {
+    //     "key": "aa0ceb",
+    //     "firstName": this.props.user.username
+    // };
 
-    let ldclient = ldClient.initialize('', user);
+    // let ldclient = ldClient.initialize('', user);
 
-    const that = this;
+    // const that = this;
 
-    ldclient.on('ready', function() {
-      let showFeature = ldclient.variation("header-color");
-      that.setState({headerBackgroundColor: `${showFeature}`});
-    });
+    // ldclient.on('ready', function() {
+    //   let showFeature = ldclient.variation("header-color");
+    //   that.setState({headerBackgroundColor: `${showFeature}`});
+    // });
   }
 
 
